+++ resolwe/flow/tests/test_annotations.py
@@ -471,6 +471,13 @@ class AnnotationViewSetsTest(TestCase):
             },
         )
 
+        # Filter by id.
+        request = factory.get("/", {"id": self.annotation_field1.pk}, format="json")
+        response = self.annotationfield_viewset(request)
+        self.assertEqual(response.status_code, status.HTTP_200_OK)
+        self.assertEqual(len(response.data), 1)
+        self.assertEqual(response.data[0]["name"], self.annotation_field1.name)
+
         # Filter by required.
         request = factory.get("/", {"required": False}, format="json")
         response = self.annotationfield_viewset(request)